pythia_bknd/requirements.txt
antiorm==1.2.1
asgiref==3.7.2
db==0.1.1
db-sqlite3==0.0.1
Django==5.0
django-cors-headers==4.3.1
django-environ==0.11.2
djangorestframework==3.14.0
mysql==0.0.3
mysqlclient==2.2.0
pytz==2023.3.post1
sqlparse==0.4.4
tzdata==2023.3
